Restructure parser package. Add python for formatter
