+++ bin/env-var-docs-create-venv
@@ -4,7 +4,14 @@
 
 source bin/lib.sh
 
-if [[ "$(which python3)" == "" ]]; then
+python_path="$(which python3)"
+
+if [[ "${python_path}" =~ "env-var-docs/venv/bin/python3" ]]; then
+  echo "ERROR: venv is currently activated, first run 'deactivate' in your shell then re-run this script."
+  exit 1
+fi
+
+if [[ "${python_path}" == "" ]]; then
   echo "System installation of python 3.10+ is required"
   exit 1
 fi
@@ -18,9 +25,18 @@ else:
     exit("System installation of python >= 3.10 is required")
 EOF
 
-if [[ ! -d env-var-docs/venv ]]; then
-  echo "venv does not exist in env-var-docs, creating..."
-  python3 -m venv env-var-docs/venv
+if [[ -d env-var-docs/venv ]]; then
+  echo "Found venv directory, removing..."
+  rm -rf env-var-docs/venv
 fi
 
-echo "To activate the virtual environment, run 'source env-var-docs/venv/bin/activate'. To deactivate the environment, run 'deactivate'"
+echo "Creating venv directory..."
+python3 -m venv env-var-docs/venv
+
+echo "Installing dependencies..."
+source env-var-docs/venv/bin/activate
+
+pip install pytest-cov mypy
+pip install -e env-var-docs/parser-package
+
+echo -e "\nIMPORTANT: To activate the virtual environment in your shell, you MUST run 'source env-var-docs/venv/bin/activate'. To deactivate the environment, run 'deactivate'."
