+++ bin/env-var-docs-run-tests
@@ -1,0 +1,20 @@
+#! /usr/bin/env bash
+
+# DOC: Runs mypy and pytest on env-var-docs directory.
+
+source bin/lib.sh
+
+if [[ ! -d env-var-docs/venv ]]; then
+  echo "ERROR: venv directory does not exist. First run 'bin/env-var-docs-create-venv' then run this script."
+  exit 1
+fi
+
+source env-var-docs/venv/bin/activate
+
+echo "Running mypy..."
+mypy env-var-docs
+
+echo "Running pytest..."
+pytest --cov=env-var-docs --cov-report=html
+
+echo "To see coverage report, open 'htmlcov/index.html' in a web browser."
