+++ env-var-docs/parser-package/env_var_docs.egg-info/SOURCES.txt
@@ -1,0 +1,7 @@
+README.md
+pyproject.toml
+setup.py
+env_var_docs.egg-info/PKG-INFO
+env_var_docs.egg-info/SOURCES.txt
+env_var_docs.egg-info/dependency_links.txt
+env_var_docs.egg-info/top_level.txt
