+++ env-var-docs/parser-package/src/env_var_docs.egg-info/SOURCES.txt
@@ -1,0 +1,8 @@
+README.md
+pyproject.toml
+src/env_var_docs/__init__.py
+src/env_var_docs/parser.py
+src/env_var_docs.egg-info/PKG-INFO
+src/env_var_docs.egg-info/SOURCES.txt
+src/env_var_docs.egg-info/dependency_links.txt
+src/env_var_docs.egg-info/top_level.txt
